+++ libft.h
@@ -20,7 +20,7 @@ typedef struct s_list
 	struct s_list	*next;
 }					t_list;
 int		ft_atoi_base(const char *nptr, const char *base);
-int		ft_atoi(const char *nptr);
+long	ft_atoi(const char *nptr);
 void	ft_bzero(void *s, size_t n);
 void	*ft_calloc(size_t nmemb, size_t size);
 int		ft_isalnum(int arg);
